refactor(#41): 퀴즈 진행과 관련된 부분은 QuizRoomProgressService로 분리하는 식으로 리팩토링
