+++ src/main/kotlin/com/tuk/oriddle/domain/quizroom/controller/QuizRoomWsController.kt
@@ -1,6 +1,7 @@
 package com.tuk.oriddle.domain.quizroom.controller
 
 import com.tuk.oriddle.domain.quizroom.dto.message.CheckAnswerMessage
+import com.tuk.oriddle.domain.quizroom.service.QuizRoomProgressService
 import com.tuk.oriddle.domain.quizroom.service.QuizRoomService
 import com.tuk.oriddle.global.result.ResultCode.*
 import org.springframework.messaging.handler.annotation.DestinationVariable
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.*
 
 @Controller
-class QuizRoomWsController(private val quizRoomService: QuizRoomService) {
+class QuizRoomWsController(private val quizRoomProgressService: QuizRoomProgressService) {
     @MessageMapping("/quiz-room/{quizRoomId}/check-answer")
     fun handleMessage(
         @DestinationVariable("quizRoomId") quizRoomId: Long,
@@ -19,6 +20,6 @@ class QuizRoomWsController(private val quizRoomService: QuizRoomService) {
         headerAccessor: SimpMessageHeaderAccessor
     ) {
         val userId = (headerAccessor.sessionAttributes?.get("id") as String).toLong()
-        quizRoomService.checkAnswer(quizRoomId, message, userId)
+        quizRoomProgressService.checkAnswer(quizRoomId, message, userId)
     }
 }
